requirements.txt
beautifulsoup4>=4.9.0
requests>=2.25.0
feedparser>=6.0.0
tqdm>=4.60.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
playwright>=1.20.0 